refactored test location
